vici: Honor optionally passed IKE configuration name in uninstall
Currently, this doesn't work for shunts.  While the associated connection
may be considered when installing the policy, the shunt manager currently
only takes the name of a child config.  So if multiple shunts under
different connections share the same name they may be installed
individually, but while uninstalling it's not possible to control which
one of them is removed if both are installed.