proposal: Remove RFC5114 MODP DH groups from default proposal
Recent research demonstrates that at least for 1024-bit DH groups, it is
possible to create specially crafted primes having a backdoor. From the
prime itself this is not detectable, creating a perferct NOBUS attach.

http://eprint.iacr.org/2016/961

The primes defined in RFC5114 do not provide any information how these primes
have been selected. In the default proposal we included one of the 2048-bit
primes only, where it is questionable if constructing a backdoored prime is
feasible. Nonetheless removes this patch the group from the set of default
proposals as well.